webrtc-quic/RTCQuicTransport.https.html

<!doctype html>
<meta charset=utf-8>
<title>RTCQuicTransport.https.html</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../webrtc/RTCIceTransport-extension-helper.js"></script>
<script src="RTCQuicTransport-helper.js"></script>
<script>
'use strict';

// These tests are based on the following specification:
// https://w3c.github.io/webrtc-quic/

// The following helper functions are called from
// RTCIceTransport-extension-helper.js:
//   makeIceTransport
//   makeAndGatherTwoIceTransports

// The following helper functions are called from RTCQuicTransport-helper.js:
//   makeQuicTransport
//   makeStandaloneQuicTransport
//   makeAndStartTwoQuicTransports
//   makeTwoConnectedQuicTransports

promise_test(async t => {
  const iceTransport = makeIceTransport(t);
  const quicTransport = makeQuicTransport(t, iceTransport);
  assert_equals(quicTransport.transport, iceTransport,
      'Expect transport to be the same as the one passed in the constructor.');
  assert_equals(quicTransport.state, 'new', `Expect state to be 'new'.`);
}, 'RTCQuicTransport initial properties are set.');

promise_test(async t => {
  const iceTransport = makeIceTransport(t);
  iceTransport.stop();
  assert_throws('InvalidStateError',
      () => makeQuicTransport(t, iceTransport));
}, 'RTCQuicTransport constructor throws if passed a closed RTCIceTransport.');

promise_test(async t => {
  const iceTransport = makeIceTransport(t);
  const firstQuicTransport =
      makeQuicTransport(t, iceTransport);
  assert_throws('InvalidStateError',
      () => makeQuicTransport(t, iceTransport));
}, 'RTCQuicTransport constructor throws if passed an RTCIceTransport that ' +
    'already has an active RTCQuicTransport.');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.stop();
  assert_equals(quicTransport.state, 'closed');
}, `stop() changes state to 'closed'.`);

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.transport.stop();
  assert_equals(quicTransport.state, 'closed');
}, `RTCIceTransport.stop() changes RTCQuicTransport.state to 'closed'.`);

promise_test(async t => {
  const [ localQuicTransport, remoteQuicTransport ] =
      await makeAndStartTwoQuicTransports(t);
  const localWatcher = new EventWatcher(t, localQuicTransport, 'statechange');
  const remoteWatcher = new EventWatcher(t, remoteQuicTransport, 'statechange');
  await Promise.all([
    localWatcher.wait_for('statechange').then(() => {
      assert_equals(localQuicTransport.state, 'connected');
    }),
    remoteWatcher.wait_for('statechange').then(() => {
      assert_equals(remoteQuicTransport.state, 'connected');
    }),
  ]);
}, 'Two RTCQuicTransports connect to each other.');

promise_test(async t => {
  const [ localQuicTransport, remoteQuicTransport ] =
      await makeTwoConnectedQuicTransports(t);
  localQuicTransport.stop();
  const remoteWatcher = new EventWatcher(t, remoteQuicTransport, 'statechange');
  await remoteWatcher.wait_for('statechange');
  assert_equals(remoteQuicTransport.state, 'closed');
}, `stop() fires a statechange event to 'closed' on the remote transport`);

promise_test(async t => {
  const localQuicTransport = await makeStandaloneQuicTransport(t);
  const remoteQuicTransport = await makeStandaloneQuicTransport(t);
  const mistmatchedKey = new Uint8Array([12345]);
  assert_not_equals(mistmatchedKey, remoteQuicTransport.key);
  localQuicTransport.listen(mistmatchedKey);
  remoteQuicTransport.connect();

  const localWatcher = new EventWatcher(t, localQuicTransport, 'statechange');
  const remoteWatcher = new EventWatcher(t, remoteQuicTransport, 'statechange');
  await Promise.all([
    localWatcher.wait_for('statechange').then(() => {
      assert_equals(localQuicTransport.state, 'failed')
    }),
    remoteWatcher.wait_for('statechange').then(() => {
      assert_equals(remoteQuicTransport.state, 'failed')
    }),
  ]);
}, 'Connection failure when keys mismatch.');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.connect();
  assert_equals(quicTransport.state, 'connecting');
}, `connect() changes state to 'connecting'.`);

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.connect();
  assert_throws('InvalidStateError',
      () => quicTransport.connect());
}, 'connect() throws if already called connect().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.listen(new Uint8Array([12345]));
  assert_throws('InvalidStateError',
      () => quicTransport.connect());
}, 'connect() throws if already called listen().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.stop();
  assert_throws('InvalidStateError',
      () => quicTransport.connect());
}, 'connect() throws after stop().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.transport.stop();
  assert_throws('InvalidStateError',
      () => quicTransport.connect());
}, 'connect() throws if called RTCIceTransport has stopped.');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.listen(new Uint8Array([12345]));
  assert_equals(quicTransport.state, 'connecting');
}, `listen() changes state to 'connecting'.`);

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.connect();
  assert_throws('InvalidStateError',
      () => quicTransport.listen(new Uint8Array([12345])));
}, 'listen() throws if already called connect().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.listen(new Uint8Array([12345]));
  assert_throws('InvalidStateError',
      () => quicTransport.listen(new Uint8Array([12345])));
}, 'listen() throws if already called listen().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.stop();
  assert_throws('InvalidStateError',
      () => quicTransport.listen(new Uint8Array([12345])));
}, 'listen() throws after stop().');

promise_test(async t => {
  const quicTransport = await makeStandaloneQuicTransport(t);
  quicTransport.transport.stop();
  assert_throws('InvalidStateError',
      () => quicTransport.listen(new Uint8Array([12345])));
}, 'listen() throws if called RTCIceTransport has stopped.');

</script>
